Code refactoring for consistent naming
Summary: Renamed the code and fixed broken compilation links

cache_tier* : Tiered cache implementation framework
blockcache* : Components of block cache implemetation
  blockcache_file : Cache file implementation
  blockcache_file_writer : Pipelined writer
  blockcache_metadata : In memory blockcache metadata
cache_volatile* : In memory cache implementation
hash_table* : Hash table implementation that metadata uses
  hash_table_evictable* : Hash table implementation where items can only be
                          evicted (not erased)
lrulist : In place update list

Test Plan:

Reviewers:

CC: leveldb@

Task ID: #8247246

Blame Rev:

+++ cache/blockcache_test.cc
@@ -10,7 +10,7 @@
 #include <iostream>
 
 #include "include/rocksdb/cache.h"
-#include "cache/blkcache.h"
+#include "cache/blockcache.h"
 #include "cache/cache_volatile.h"
 #include "util/testharness.h"
 #include "util/arena.h"
@@ -139,7 +139,7 @@ class BlockCacheImplTest : public testing::Test {
     assert(pad_size);
 
     char ret[pad_size];
-    size_t pos = pad_size - 1;
+    int pos = static_cast<int>(pad_size) - 1;
     size_t count = 0;
     size_t t = data;
 
@@ -173,7 +173,7 @@ class BlockCacheImplTest : public testing::Test {
       auto k = prefix + PaddedNumber(i, /*count=*/ 8);
       Slice key(k);
       unique_ptr<char[]> block;
-      uint32_t block_size;
+      size_t block_size;
       if (relaxed && !cache_->Lookup(key, &block, &block_size)) {
         continue;
       }
